Moved EnumIntegerSet to package conv
Moved EnumIntegerSet to package conv
Moved the appropiate Test
Updated imports in all accessors
