+++ src/main/java/dev/dokan/dokan_java/constants/microsoft/FileAttribute.java
@@ -1,7 +1,7 @@
 package dev.dokan.dokan_java.constants.microsoft;
 
 import dev.dokan.dokan_java.conv.EnumInteger;
-import dev.dokan.dokan_java.structure.EnumIntegerSet;
+import dev.dokan.dokan_java.conv.EnumIntegerSet;
 import com.sun.jna.platform.win32.WinNT;
 
 /**
