+++ src/main/java/dev/dokan/dokan_java/constants/microsoft/FileSystemFlag.java
@@ -2,7 +2,7 @@ package dev.dokan.dokan_java.constants.microsoft;
 
 import dev.dokan.dokan_java.DokanOperations;
 import dev.dokan.dokan_java.conv.EnumInteger;
-import dev.dokan.dokan_java.structure.EnumIntegerSet;
+import dev.dokan.dokan_java.conv.EnumIntegerSet;
 import com.sun.jna.platform.win32.WinNT;
 
 /**
